+++ package.json
@@ -5,12 +5,10 @@
   "main": "index.js",
   "scripts": {
     "build": "run-s build:*",
-    "build:babel": "babel src/ --out-dir lib/ --source-maps inline",
     "build:browserify": "browserify --verbose -t [ babelify ] sample/source.js -o sample/bundle.js",
     "test": "run-s test:*",
     "test:mocha": "mocha 'test/*.js' -r babel-polyfill --compilers js:babel-register",
     "test:standard": "standard",
-    "clean": "rm -rf lib/",
     "push-gh-pages": "git status -s | grep '.' && echo 'you should commit before push-gh-pages' || (git add -f ./sample && git commit -m 'push gh-pages' && git push -f origin HEAD:gh-pages; git reset --hard HEAD^)"
   },
   "author": "Sho Hashimoto <hashimoto@shokai.org>",
